requirements.txt
easyocr
pymupdf
pillow
opencv-python
python-bidi==0.4.
langchain-community
langchain
langchain-groq
python-dotenv
langchain_ollama
matplotlib
bs4
tqdm
unstructured
pandas
openai
tiktoken
faiss-cpu
PyPDF2